Inline show toast method. Add context where needed
