+++ app/src/main/java/mozilla/lockbox/model/ToastNotificationViewModel.kt
@@ -12,5 +12,5 @@ import androidx.annotation.StringRes
 data class ToastNotificationViewModel(
     @StringRes val strId: Int,
     @DrawableRes val img: Int,
-    val deletedItem: String? = null
+    val untranslatableString: String? = null
 )
