+++ app/src/main/java/mozilla/lockbox/presenter/AppRoutePresenter.kt
@@ -6,6 +6,7 @@
 
 package mozilla.lockbox.presenter
 
+import android.content.Context
 import android.os.Bundle
 import androidx.annotation.IdRes
 import androidx.appcompat.app.AppCompatActivity
@@ -29,10 +30,11 @@ import mozilla.lockbox.view.FingerprintAuthDialogFragment
 @ExperimentalCoroutinesApi
 class AppRoutePresenter(
     private val activity: AppCompatActivity,
+    context: Context,
     dispatcher: Dispatcher = Dispatcher.shared,
     private val routeStore: RouteStore = RouteStore.shared,
     private val settingStore: SettingStore = SettingStore.shared
-) : RoutePresenter(activity, dispatcher, routeStore) {
+) : RoutePresenter(activity, context, dispatcher, routeStore) {
 
     override fun onViewReady() {
         super.onViewReady()
