+++ app/src/main/java/mozilla/lockbox/presenter/AutofillRoutePresenter.kt
@@ -1,6 +1,7 @@
 package mozilla.lockbox.presenter
 
 import android.app.Activity
+import android.content.Context
 import android.content.Intent
 import android.os.Build
 import android.service.autofill.FillResponse
@@ -33,6 +34,7 @@ import mozilla.lockbox.view.FingerprintAuthDialogFragment
 @ExperimentalCoroutinesApi
 @RequiresApi(Build.VERSION_CODES.O)
 open class AutofillRoutePresenter(
+    context: Context,
     private val activity: AppCompatActivity,
     private val responseBuilder: FillResponseBuilder,
     private val dispatcher: Dispatcher = Dispatcher.shared,
@@ -40,7 +42,7 @@ open class AutofillRoutePresenter(
     private val autofillStore: AutofillStore = AutofillStore.shared,
     private val dataStore: DataStore = DataStore.shared,
     private val pslSupport: PublicSuffixSupport = PublicSuffixSupport.shared
-) : RoutePresenter(activity, dispatcher, routeStore) {
+) : RoutePresenter(activity, context, dispatcher, routeStore) {
 
     override fun onViewReady() {
         navController = Navigation.findNavController(activity, R.id.autofill_fragment_nav_host)
