+++ app/src/main/java/mozilla/lockbox/view/AutofillRootActivity.kt
@@ -23,7 +23,7 @@ class AutofillRootActivity : AppCompatActivity() {
 
         // pull responsebuilder & search status out of intent at a later iteration of this
         val responseBuilder = IntentBuilder.getResponseBuilder(intent)
-        presenter = AutofillRoutePresenter(this, responseBuilder)
+        presenter = AutofillRoutePresenter(context = , this, responseBuilder)
         presenter.onViewReady()
     }
