+++ app/src/test/java/mozilla/lockbox/presenter/AppRoutePresenterTest.kt
@@ -85,6 +85,7 @@ class AppRoutePresenterTest {
 
         subject = AppRoutePresenter(
             activity,
+            context,
             dispatcher,
             routeStore,
             settingStore
